fix #4

+++ tritonus-remaining/src/main/java/org/tritonus/sampled/convert/SampleRateConversionProvider.java
@@ -674,7 +674,7 @@ public class SampleRateConversionProvider extends
 										+ " history length=" + historyLength);
 							}
 						} else {
-							value += inSamples[iInIndex] * factor;
+                            value += (float) (inSamples[iInIndex] * factor);
 						}
 						factor = 1 - factor;
 						iInIndex++;
